requirements-dev.txt
black>=23.0.0
flake8>=6.0.0
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
mypy>=1.0.0
safety>=2.0.0
bandit>=1.7.0
codecov>=2.1.0